refactor: Result클래스 삭제 및 책임 이동
Round가 충분히 가져도 될 책임이라 생각하여 result를 삭제 후 Round로 변경함
